Lock users after x failed attempts
After a configurable amount of consecutive failed login attempts the new attribute `locked_until` will be set.
The default amount of attempts is 5 and the default time the user is locked is 5 minutes.
After this commit locked users will still be able to log in. Locking out locked users is part of the next
commit.