Fixes the check whether a user is actually locked.
Before, the user was claimed to be locked when any CASino::User sharing
the same username was locked. But a user is actually only locked, if he
does not have any CASino::User left, which is unlocked. The reason
herefore is, that he could have any secondary/legacy authentication
providers, on which he can never login. Therefore, one of his 'alter
egos' may be getting locked on any regular (and otherwise successful)
login attempt and he would be locked out regardless. That is now
changed, so that all CASino::Users sharing the same username must be
first be in locked state before one is considered locked.